+++ auto_route/example/android/app/src/main/kotlin/com/ma/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.ma.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}