+++ auto_route/lib/src/router/auto_route_page.dart
@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 
 abstract class AutoRoutePage<T> extends Page<T> {
   final RouteData routeData;
-  final Widget child;
+  final Widget _child;
   final bool fullscreenDialog;
   final bool maintainState;
 
@@ -18,11 +18,16 @@ abstract class AutoRoutePage<T> extends Page<T> {
 
   AutoRoutePage({
     required this.routeData,
-    required this.child,
+    required Widget child,
     this.fullscreenDialog = false,
     this.maintainState = true,
     LocalKey? key,
-  }) : super(
+  })  : _child = child is AutoRouteWrapper
+            ? WrappedRoute(
+                child: child as AutoRouteWrapper,
+              )
+            : child,
+        super(
           restorationId: routeData.name,
           name: routeData.name,
           arguments: routeData.route.args,
@@ -30,14 +35,15 @@ abstract class AutoRoutePage<T> extends Page<T> {
 
   @override
   bool canUpdate(Page<dynamic> other) {
-    return other.runtimeType == runtimeType && (other as AutoRoutePage).routeKey == routeKey;
+    return other.runtimeType == runtimeType &&
+        (other as AutoRoutePage).routeKey == routeKey;
   }
 
   LocalKey get routeKey => routeData.key;
 
   Widget buildPage(BuildContext context) {
     return RouteDataScope(
-      child: child,
+      child: _child,
       routeData: routeData,
     );
   }
@@ -74,7 +80,8 @@ class MaterialPageX<T> extends AutoRoutePage<T> {
   }
 }
 
-class PageBasedMaterialPageRoute<T> extends PageRoute<T> with MaterialRouteTransitionMixin<T> {
+class PageBasedMaterialPageRoute<T> extends PageRoute<T>
+    with MaterialRouteTransitionMixin<T> {
   PageBasedMaterialPageRoute({
     required AutoRoutePage page,
   }) : super(settings: page);
@@ -96,7 +103,8 @@ class PageBasedMaterialPageRoute<T> extends PageRoute<T> with MaterialRouteTrans
   String get debugLabel => '${super.debugLabel}(${_page.name})';
 }
 
-class _CustomPageBasedPageRouteBuilder<T> extends PageRoute<T> with _CustomPageRouteTransitionMixin<T> {
+class _CustomPageBasedPageRouteBuilder<T> extends PageRoute<T>
+    with _CustomPageRouteTransitionMixin<T> {
   _CustomPageBasedPageRouteBuilder({
     required AutoRoutePage page,
   }) : super(settings: page);
@@ -114,7 +122,8 @@ class _CustomPageBasedPageRouteBuilder<T> extends PageRoute<T> with _CustomPageR
   String get debugLabel => '${super.debugLabel}(${_page.name})';
 }
 
-class _NoAnimationPageRouteBuilder<T> extends PageRoute<T> with _NoAnimationPageRouteTransitionMixin<T> {
+class _NoAnimationPageRouteBuilder<T> extends PageRoute<T>
+    with _NoAnimationPageRouteTransitionMixin<T> {
   _NoAnimationPageRouteBuilder({
     required AutoRoutePage page,
   }) : super(settings: page);
@@ -157,10 +166,14 @@ mixin _NoAnimationPageRouteTransitionMixin<T> on PageRoute<T> {
   @override
   bool canTransitionTo(TransitionRoute<dynamic> nextRoute) {
     // Don't perform outgoing animation if the next route is a fullscreen dialog.
-    return (nextRoute is _CustomPageBasedPageRouteBuilder && !nextRoute.fullscreenDialog ||
-            nextRoute is MaterialRouteTransitionMixin && !nextRoute.fullscreenDialog) ||
-        (nextRoute is _NoAnimationPageRouteTransitionMixin && !nextRoute.fullscreenDialog) ||
-        (nextRoute is CupertinoRouteTransitionMixin && !nextRoute.fullscreenDialog);
+    return (nextRoute is _CustomPageBasedPageRouteBuilder &&
+                !nextRoute.fullscreenDialog ||
+            nextRoute is MaterialRouteTransitionMixin &&
+                !nextRoute.fullscreenDialog) ||
+        (nextRoute is _NoAnimationPageRouteTransitionMixin &&
+            !nextRoute.fullscreenDialog) ||
+        (nextRoute is CupertinoRouteTransitionMixin &&
+            !nextRoute.fullscreenDialog);
   }
 
   @override
@@ -198,7 +211,8 @@ mixin _CustomPageRouteTransitionMixin<T> on PageRoute<T> {
   bool get barrierDismissible => _page.barrierDismissible;
 
   @override
-  Color? get barrierColor => _page.barrierColor == null ? null : Color(_page.barrierColor!);
+  Color? get barrierColor =>
+      _page.barrierColor == null ? null : Color(_page.barrierColor!);
 
   @override
   String? get barrierLabel => _page.barrierLabel;
@@ -209,9 +223,12 @@ mixin _CustomPageRouteTransitionMixin<T> on PageRoute<T> {
   @override
   bool canTransitionTo(TransitionRoute<dynamic> nextRoute) {
     // Don't perform outgoing animation if the next route is a fullscreen dialog.
-    return (nextRoute is MaterialRouteTransitionMixin && !nextRoute.fullscreenDialog) ||
-        (nextRoute is _NoAnimationPageRouteTransitionMixin && !nextRoute.fullscreenDialog) ||
-        (nextRoute is CupertinoRouteTransitionMixin && !nextRoute.fullscreenDialog);
+    return (nextRoute is MaterialRouteTransitionMixin &&
+            !nextRoute.fullscreenDialog) ||
+        (nextRoute is _NoAnimationPageRouteTransitionMixin &&
+            !nextRoute.fullscreenDialog) ||
+        (nextRoute is CupertinoRouteTransitionMixin &&
+            !nextRoute.fullscreenDialog);
   }
 
   @override
@@ -227,13 +244,19 @@ mixin _CustomPageRouteTransitionMixin<T> on PageRoute<T> {
     );
   }
 
-  Widget _defaultTransitionsBuilder(BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation, Widget child) {
+  Widget _defaultTransitionsBuilder(
+      BuildContext context,
+      Animation<double> animation,
+      Animation<double> secondaryAnimation,
+      Widget child) {
     return child;
   }
 
   @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation, Widget child) {
-    final transitionsBuilder = _page.transitionsBuilder ?? _defaultTransitionsBuilder;
+  Widget buildTransitions(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    final transitionsBuilder =
+        _page.transitionsBuilder ?? _defaultTransitionsBuilder;
     return transitionsBuilder(context, animation, secondaryAnimation, child);
   }
 }
@@ -276,7 +299,8 @@ class CupertinoPageX<T> extends _TitledAutoRoutePage<T> {
   }
 }
 
-class _PageBasedCupertinoPageRoute<T> extends PageRoute<T> with CustomCupertinoRouteTransitionMixin<T> {
+class _PageBasedCupertinoPageRoute<T> extends PageRoute<T>
+    with CustomCupertinoRouteTransitionMixin<T> {
   _PageBasedCupertinoPageRoute({
     required _TitledAutoRoutePage page,
   }) : super(settings: page);
@@ -324,7 +348,8 @@ class AdaptivePage<T> extends _TitledAutoRoutePage<T> {
   }
 }
 
-typedef CustomRouteBuilder = Route<T> Function<T>(BuildContext context, Widget child, CustomPage<T> page);
+typedef CustomRouteBuilder = Route<T> Function<T>(
+    BuildContext context, Widget child, CustomPage<T> page);
 
 class CustomPage<T> extends AutoRoutePage<T> {
   final bool opaque;