+++ auto_route/test/nested_tabs_router/router.dart
@@ -13,6 +13,7 @@ part 'router.gr.dart';
       path: '/',
       page: TabsHostPage,
       children: tabRoutes,
+
     ),
     RedirectRoute(path: '*', redirectTo: '/'),
   ],