+++ auto_route_generator/lib/src/resolvers/route_config_resolver.dart
@@ -29,7 +29,7 @@ class RouteConfigResolver {
 
   RouteConfigResolver(this._routerConfig, this._typeResolver);
 
-  RouteConfig resolve(ConstantReader autoRoute) {
+  RouteConfig resolve(ConstantReader autoRoute, List<PathParamConfig> inheritedPathParams) {
     final page = autoRoute.peek('page')?.typeValue;
     var path = autoRoute.peek('path')?.stringValue;
     if (page == null) {
@@ -55,8 +55,7 @@ class RouteConfigResolver {
     );
 
     final classElement = page.element as ClassElement;
-    final hasWrappedRoute = classElement.allSupertypes.any((e) =>
-        e.getDisplayString(withNullability: false) == 'AutoRouteWrapper');
+    final hasWrappedRoute = classElement.allSupertypes.any((e) => e.getDisplayString(withNullability: false) == 'AutoRouteWrapper');
     var pageType = _typeResolver.resolveType(page);
     var className = page.getDisplayString(withNullability: false);
 
@@ -69,27 +68,25 @@ class RouteConfigResolver {
       }
     }
     throwIf(
-      path.startsWith("/") && _routerConfig.parent != null,
+      path.startsWith('/') && _routerConfig.parent != null,
       'Child [$path] can not start with a forward slash',
     );
 
     var pathName = path;
     var pathParams = RouteParameterResolver.extractPathParams(path);
 
-    var fullscreenDialog = autoRoute.peek('fullscreenDialog')?.boolValue;
-    var maintainState = autoRoute.peek('maintainState')?.boolValue;
-    var fullMatch = autoRoute.peek('fullMatch')?.boolValue;
-    var initial = autoRoute.peek('initial')?.boolValue ?? false;
-    var usesPathAsKey = autoRoute.peek('usesPathAsKey')?.boolValue ?? false;
+    final fullscreenDialog = autoRoute.peek('fullscreenDialog')?.boolValue;
+    final maintainState = autoRoute.peek('maintainState')?.boolValue;
+    final fullMatch = autoRoute.peek('fullMatch')?.boolValue;
+    final initial = autoRoute.peek('initial')?.boolValue ?? false;
+    final usesPathAsKey = autoRoute.peek('usesPathAsKey')?.boolValue ?? false;
 
     var guards = <ResolvedType>[];
-    autoRoute
-        .peek('guards')
-        ?.listValue
-        .map((g) => g.toTypeValue())
-        .forEach((guard) {
-      guards.add(_typeResolver.resolveType(guard!));
-    });
+    autoRoute.peek('guards')?.listValue.map((g) => g.toTypeValue()).forEach(
+      (guard) {
+        guards.add(_typeResolver.resolveType(guard!));
+      },
+    );
 
     var returnType = ResolvedType(name: 'dynamic');
     var dartType = autoRoute.objectValue.type;
@@ -117,50 +114,38 @@ class RouteConfigResolver {
       cupertinoNavTitle = autoRoute.peek('cupertinoPageTitle')?.stringValue;
     } else if (autoRoute.instanceOf(TypeChecker.fromRuntime(CustomRoute))) {
       routeType = RouteType.custom;
-      durationInMilliseconds =
-          autoRoute.peek('durationInMilliseconds')?.intValue;
-      reverseDurationInMilliseconds =
-          autoRoute.peek('reverseDurationInMilliseconds')?.intValue;
+      durationInMilliseconds = autoRoute.peek('durationInMilliseconds')?.intValue;
+      reverseDurationInMilliseconds = autoRoute.peek('reverseDurationInMilliseconds')?.intValue;
       customRouteOpaque = autoRoute.peek('opaque')?.boolValue;
-      customRouteBarrierDismissible =
-          autoRoute.peek('barrierDismissible')?.boolValue;
+      customRouteBarrierDismissible = autoRoute.peek('barrierDismissible')?.boolValue;
       customRouteBarrierLabel = autoRoute.peek('barrierLabel')?.stringValue;
-      final function =
-          autoRoute.peek('transitionsBuilder')?.objectValue.toFunctionValue();
+      final function = autoRoute.peek('transitionsBuilder')?.objectValue.toFunctionValue();
       if (function != null) {
         transitionBuilder = _typeResolver.resolveFunctionType(function);
       }
-      final builderFunction =
-          autoRoute.peek('customRouteBuilder')?.objectValue.toFunctionValue();
+      final builderFunction = autoRoute.peek('customRouteBuilder')?.objectValue.toFunctionValue();
       if (builderFunction != null) {
         customRouteBuilder = _typeResolver.resolveFunctionType(builderFunction);
       }
       customRouteBarrierColor = autoRoute.peek('barrierColor')?.intValue;
     } else {
-      var globConfig = _routerConfig.globalRouteConfig;
+      var globConfig = _routerConfig.parentRouteConfig;
       routeType = globConfig.routeType;
       if (globConfig.routeType == RouteType.custom) {
         transitionBuilder = globConfig.transitionBuilder;
         durationInMilliseconds = globConfig.durationInMilliseconds;
-        customRouteBarrierDismissible =
-            globConfig.customRouteBarrierDismissible;
+        customRouteBarrierDismissible = globConfig.customRouteBarrierDismissible;
         customRouteOpaque = globConfig.customRouteOpaque;
-        reverseDurationInMilliseconds =
-            globConfig.reverseDurationInMilliseconds;
+        reverseDurationInMilliseconds = globConfig.reverseDurationInMilliseconds;
         customRouteBuilder = globConfig.customRouteBuilder;
       }
     }
 
     final meta = <MetaEntry>[];
-    for (final entry in autoRoute
-        .read('meta')
-        .mapValue
-        .entries
-        .where((e) => e.value?.type != null)) {
-      final valueType =
-          entry.value!.type!.getDisplayString(withNullability: false);
-      throwIf(!validMetaValues.contains(valueType),
-          'Meta value type $valueType is not supported!\nSupported types are $validMetaValues');
+    for (final entry in autoRoute.read('meta').mapValue.entries.where((e) => e.value?.type != null)) {
+      final valueType = entry.value!.type!.getDisplayString(withNullability: false);
+      throwIf(
+          !validMetaValues.contains(valueType), 'Meta value type ${valueType} is not supported!\nSupported types are ${validMetaValues}');
       switch (valueType) {
         case 'bool':
           {
@@ -214,14 +199,16 @@ class RouteConfigResolver {
     var params = constructor!.parameters;
     var parameters = <ParamConfig>[];
     if (params.isNotEmpty == true) {
-      if (constructor.isConst &&
-          params.length == 1 &&
-          params.first.type.getDisplayString(withNullability: false) == 'Key') {
+      if (constructor.isConst && params.length == 1 && params.first.type.getDisplayString(withNullability: false) == 'Key') {
         hasConstConstructor = true;
       } else {
         final paramResolver = RouteParameterResolver(_typeResolver);
         for (ParameterElement p in constructor.parameters) {
-          parameters.add(paramResolver.resolve(p));
+          parameters.add(paramResolver.resolve(
+            p,
+            pathParams: pathParams,
+            inheritedPathParams: inheritedPathParams,
+          ));
         }
       }
     }
@@ -229,13 +216,9 @@ class RouteConfigResolver {
     var pathParameters = parameters.where((element) => element.isPathParam);
 
     if (parameters.any((p) => p.isPathParam || p.isQueryParam)) {
-      var unParsableRequiredArgs = parameters.where((p) =>
-          (p.isRequired || p.isPositional) &&
-          !p.isPathParam &&
-          !p.isQueryParam);
+      var unParsableRequiredArgs = parameters.where((p) => (p.isRequired || p.isPositional) && !p.isPathParam && !p.isQueryParam);
       if (unParsableRequiredArgs.isNotEmpty) {
-        print(
-            '\nWARNING => Because [$className] has required parameters ${unParsableRequiredArgs.map((e) => e.paramName)} '
+        print('\nWARNING => Because [$className] has required parameters ${unParsableRequiredArgs.map((e) => e.paramName)} '
             'that can not be parsed from path,\n@PathParam() and @QueryParam() annotations will be ignored.\n');
       }
     }